Move Reloader cog to directory.
